@@ -13,19 +13,18 @@
     "author": "",
     "license": "Apache-2.0",
     "devDependencies": {
-        "@eslint/js": "^9.13.0",
-        "@tsconfig/node18": "^18.2.4",
+        "@eslint/js": "^9.25.1",
         "@types/jest": "^29.5.14",
-        "@types/node": "^18.19.62",
-        "eslint": "^9.13.0",
-        "eslint-config-prettier": "^9.1.0",
-        "eslint-plugin-jest": "^28.8.3",
+        "@types/node": "^20.17.31",
+        "eslint": "^9.25.1",
+        "eslint-config-prettier": "^10.1.2",
+        "eslint-plugin-jest": "^28.11.0",
         "fabric-network": "^2.2.20",
         "fabric-wallet-migration": "file:../fabric-wallet-migration-dev.tgz",
         "jest": "^29.7.0",
-        "prettier": "^3.3.3",
-        "ts-jest": "^29.2.5",
-        "typescript": "~5.6.0",
-        "typescript-eslint": "^8.12.2"
+        "prettier": "^3.5.3",
+        "ts-jest": "^29.3.2",
+        "typescript": "~5.8.3",
+        "typescript-eslint": "^8.31.0"
     }
 }
