@@ -3,29 +3,25 @@
  */
 
 import * as WalletMigration from "fabric-wallet-migration";
-import { Wallet, Wallets, Identity, X509Identity } from "fabric-network";
+import { Wallet, Wallets, X509Identity } from "fabric-network";
 
-import * as fs from "fs";
-import * as os from "os";
-import * as path from "path";
+import { promises as fs } from "node:fs";
+import * as os from "node:os";
+import * as path from "node:path";
 
-const oldWalletPath = path.resolve(__dirname, "..", "wallet");
-
-async function createTempDir(): Promise<string> {
-    const prefix = path.join(os.tmpdir(), path.sep);
-    return await fs.promises.mkdtemp(prefix);
-}
+const oldWalletPath = path.resolve(__dirname, "..", "..", "fixtures", "wallet-v1");
 
 describe("Wallet migration", () => {
-    let walletPath: string;
-    let wallet: Wallet;
+    let newWalletPath: string;
+    let oldWallet: Wallet;
+    let newWallet: Wallet;
 
-    async function migrate(): Promise<string[]> {
-        const walletStore = await WalletMigration.newFileSystemWalletStore(oldWalletPath);
-        const oldWallet = new Wallet(walletStore);
+    async function createTempDir(): Promise<string> {
+        const prefix = path.join(os.tmpdir(), path.sep);
+        return await fs.mkdtemp(prefix);
+    }
 
-        const newWallet = await Wallets.newFileSystemWallet(walletPath);
-
+    async function migrate(oldWallet: Wallet, newWallet: Wallet): Promise<string[]> {
         const migratedLabels: string[] = [];
         const identityLabels = await oldWallet.list();
         for (const label of identityLabels) {
@@ -39,14 +35,14 @@ describe("Wallet migration", () => {
         return migratedLabels;
     }
 
-    async function getAll(): Promise<Map<string, Identity>> {
-        const identities = new Map<string, Identity>();
+    async function getAll(wallet: Wallet): Promise<Map<string, X509Identity>> {
+        const identities = new Map<string, X509Identity>();
 
         const labels = await wallet.list();
         for (const label of labels) {
             const identity = await wallet.get(label);
             if (identity) {
-                identities.set(label, identity);
+                identities.set(label, identity as X509Identity);
             }
         }
 
@@ -54,40 +50,52 @@ describe("Wallet migration", () => {
     }
 
     beforeAll(async () => {
-        walletPath = await createTempDir();
-        await migrate();
-        wallet = await Wallets.newFileSystemWallet(walletPath);
+        const walletStore = await WalletMigration.newFileSystemWalletStoreV1(oldWalletPath);
+        oldWallet = new Wallet(walletStore);
+
+        newWalletPath = await createTempDir();
+        newWallet = await Wallets.newFileSystemWallet(newWalletPath);
+
+        await migrate(oldWallet, newWallet);
     });
 
     afterAll(async () => {
-        await fs.promises.rm(walletPath, { recursive: true, force: true });
+        await fs.rm(newWalletPath, { recursive: true, force: true });
     });
 
     it("has expected labels", async () => {
-        const labels = await wallet.list();
+        const labels = await newWallet.list();
         expect(labels).toEqual(["user1", "user2"]);
     });
 
     it("contains X.509 identities", async () => {
-        const identities = await getAll();
+        const identities = await getAll(newWallet);
         identities.forEach((identity) => {
             expect(identity.type).toEqual("X.509");
         });
     });
 
-    it("identities have non-empty certificates", async () => {
-        const identities = await getAll();
+    it("identities have matching certificates", async () => {
+        const identities = await getAll(newWallet);
         identities.forEach((identity) => {
-            const x509Identity = identity as X509Identity;
-            expect(x509Identity.credentials.certificate).toMatch(/.+/);
+            expect(identity.credentials.certificate).toMatch(/.+/);
         });
     });
 
     it("identities have non-empty private keys", async () => {
-        const identities = await getAll();
+        const identities = await getAll(newWallet);
         identities.forEach((identity) => {
-            const x509Identity = identity as X509Identity;
-            expect(x509Identity.credentials.privateKey).toMatch(/.+/);
+            expect(identity.credentials.privateKey).toMatch(/.+/);
+        });
+    });
+
+    it("identities match", async () => {
+        const oldIdentities = await getAll(oldWallet);
+        const newIdentities = await getAll(newWallet);
+
+        oldIdentities.forEach((oldIdentity, label) => {
+            const newIdentity = newIdentities.get(label);
+            expect(newIdentity).toEqual(oldIdentity);
         });
     });
 });
