@@ -1,5 +1,8 @@
-/**
+/*
  * SPDX-License-Identifier: Apache-2.0
  */
 
-export { newFileSystemWalletStore } from "./WalletStores";
+export { newFileSystemWalletStoreV1, type WalletStore } from "./WalletStore";
+export { Wallet } from "./Wallet";
+export * as wallets from "./Wallets";
+export { type Identity } from "./Identity";
